# improve perf for SqlIndexedMerkleTree.update_leaf