## Change Summary
This PR significantly improves the Indexed Merkle Tree implementation and expands test coverage:

* Optimized the update_leaf method to use bulk_get_sibling_hashes for Merkle path updates
* Added `test_speed_prove_and_update` performance benchmark test


## Pros
* Improved performance for Merkle tree update operations (fetching multiple node hashes with a single query)
* Significantly reduced database access count, increasing overall throughput

## Benchmark Results

run `trees::merkle_tree::tests::test_speed_prove_and_update` test, and check execution time.
Benchmark measurements show approximately 2x speed improvement 👍 

### Before

```
running 1 test
test trees::merkle_tree::tests::test_speed_prove_and_update has been running for over 60 seconds
SqlIndexedMerkleTree.prove_and_insert: 1000 times, 32 height, 68.938687666 seconds
SqlIndexedMerkleTree.prove_and_update: 1000 times, 32 height, 33.755922417 seconds
test trees::merkle_tree::tests::test_speed_prove_and_update ... ok
```

### After

```
running 1 test
SqlIndexedMerkleTree.prove_and_insert: 1000 times, 32 height, 37.856773791 seconds
SqlIndexedMerkleTree.prove_and_update: 1000 times, 32 height, 19.517532583 seconds
test trees::merkle_tree::tests::test_speed_prove_and_update ... ok
```